Refactor Command objects
Refactor Command objects out of CLI and into their own module
