+++ src/update.rs
@@ -9,7 +9,7 @@ impl graph::Graph {
     /// Function iterates through all messages currently in the channel, performing their operations on the graph.
     pub fn client_update(
         &mut self,
-        _channel: &mpsc::Receiver<crate::cli::Command>,
+        _channel: &mpsc::Receiver<crate::command::Command>,
     ) -> anyhow::Result<()> {
         todo!()
     }
